refactor(2270): chain-proof policy gets its own module, the nonce one parse
PR-3 review round 3, mechanical half. No behaviour change; the semantics land in
the commit on top of this one.

`publisher-chain-proof.ts` takes the ~350 lines of chain-proof policy this PR grew
inside `publisher-runner.ts`: the adapter map, the resolver factory, the nonce and
identity proofs, adapter selection and the lift-recovery mapping. The runner stays the
composition root. The split is not about size — it is that the file answering "how is
the daemon assembled" had also come to answer "when may a held job be resent", and only
the second question has a double publish at the end of it. The module states the rule
once at the top: an absence must be ESTABLISHED, never inferred.

`signPopulatedTransaction` returns `{signedTx, txHash, nonce}` — a named
`SignedTransactionEnvelope` — so the signed bytes are decoded ONCE, where they are
signed and already being decoded for the hash. `nonceOfSignedTx` is gone with the
second decode it existed for, and `dispatchSerializedV10Write` takes the nonce from the
envelope rather than re-deriving it. Still fail-closed: a signer that cannot produce a
nonce leaves it absent and the recovery side keeps holding.

The real-signing rows now drive the PRODUCTION signer rather than a hand-signed
transaction, so the hash and nonce they assert on are both produced by the code under
test, cross-checked against an independent decode of the same bytes.
